file buffer: refactoring — size_t

+++ include/LunatiX/LX_FileBuffer.hpp
@@ -67,7 +67,7 @@ class LX_FileBuffer
 public:
 
     /**
-    *   @fn LX_FileBuffer(const std::string& filename, uint32_t offset = 0, uint32_t sz = 0)
+    *   @fn LX_FileBuffer(const std::string& filename, size_t offset = 0, size_t sz = 0)
     *
     *   @param [in] filename The file to read
     *   @param [in] offset The position in the file to start reading (default value: 0)
@@ -84,9 +84,9 @@ public:
     *   @exception std::logic_error If the filename is not defined
     *   @exception IOException If the file cannot be read
     */
-    LX_FileBuffer(const std::string& filename, uint32_t offset = 0, uint32_t sz = 0);
+    LX_FileBuffer(const std::string& filename, size_t offset = 0, size_t sz = 0);
     /**
-    *   @fn explicit LX_FileBuffer(const UTF8string& filename, uint32_t offset = 0, uint32_t sz = 0)
+    *   @fn explicit LX_FileBuffer(const UTF8string& filename, size_t offset = 0, size_t sz = 0)
     *
     *   @param [in] filename The file to read
     *   @param [in] offset The position in the file to start reading (default value: 0)
@@ -104,7 +104,7 @@ public:
     *   @exception IOException If the file cannot be read by the instance
     *
     */
-    explicit LX_FileBuffer(const UTF8string& filename, uint32_t offset = 0, uint32_t sz = 0);
+    explicit LX_FileBuffer(const UTF8string& filename, size_t offset = 0, size_t sz = 0);
 
     /**
     *   @fn LX_Graphics::LX_BufferedImage * loadBufferedImage(LX_Graphics::LX_PIXELFORMAT format = LX_Graphics::LX_PIXELFORMAT::RGBA8888) const
